staging: comedi: rti800: remove am9513.h header
This driver is the only file that includes the am9513.h header and
nothing from that header is actually used. Using the macros provided
by am9513.h is a bit ugly due to having to predefine a number of
additional macros to handle the low-level I/O to the Am9513 device.

Just remove the header and the unused I/O helpers in the rti800
driver. If support for the am9513 is added later it wil be implemented
properly at that time.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>
Cc: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>